Add gen columnar shuffle exchange interface
