+++ backends-velox/src/main/scala/io/glutenproject/backendsapi/velox/VeloxBackend.scala
@@ -382,7 +382,6 @@ object BackendSettings extends BackendSettingsApi {
   override def fallbackAggregateWithChild(): Boolean = true
 
   override def recreateJoinExecOnFallback(): Boolean = true
-  override def removeHashColumnFromColumnarShuffleExchangeExec(): Boolean = true
   override def rescaleDecimalLiteral(): Boolean = true
 
   /** Get the config prefix for each backend */
